engine/requirements.txt
# AI Agent 核心框架
crewai==0.28.8
crewai-tools==0.2.6

# LLM 溝通底層 (如果你用 OpenAI)
langchain-openai==0.1.7

# 環境變數與機敏金鑰管理 (強烈建議)
python-dotenv==1.0.1